# smart-ground-hub/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SmartGround API
  version: 2.0.0
  description: >
    REST API for IoT device management with MQTT-based device discovery and
    command signaling. Includes endpoints to manage device types, register
    devices, and trigger config pushes to SmartBox firmware via MQTT.

    ## Device Management

    All device operations use the flat `/api/devices` endpoint. Devices are
    always registered on a SmartBox (network access point) but are managed
    as a flat list with optional filtering by SmartBox or Range.
    Use **POST /api/devices** to register and immediately trigger a config push
    to the SmartBox firmware via MQTT.

    ## Breaking changes from v2.x

    - `POST /api/smart-boxes/{id}/devices` removed — use `POST /api/devices`
    - `GET /api/smart-boxes/{id}/devices` removed — use `GET /api/devices?smartBoxId=X`
    - `POST/DELETE /api/ranges/{rangeId}/devices/{deviceId}` removed — use `POST/DELETE /api/devices/{id}/range`
    - `SmartBoxResponse.deviceCount` removed
servers:
  - url: http://localhost:8080
    description: JVM App
  - url: http://localhost:8081
    description: Native App
security:
  - bearerAuth: []
paths:
...
  /api/device-types/{id}:
    get:
      summary: Get a device type by ID
      operationId: getDeviceType
      tags: [DeviceType]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device type found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeResponse'
        '404':
          description: Device type not found

    patch:
      summary: Update a device type
      operationId: updateDeviceType
      tags: [DeviceType]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceTypeRequest'
      responses:
        '200':
          description: Device type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeResponse'
        '404':
          description: Device type not found

  # ─────────────────────────────────────────────────────────────────────────────
  # DeviceTypeGroup (consolidated under /device-types)
  # ─────────────────────────────────────────────────────────────────────────────
...
  /api/devices:
    get:
      summary: List all devices
      operationId: listDevices
      tags: [Device]
      parameters:
        - name: rangeId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter devices by range
        - name: smartBoxId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter devices by smart box
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Zero-based page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Number of items per page
      responses:
        '200':
          description: A list of devices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePageResponse'

    post:
      summary: Register a device on a smart box
      operationId: createDevice
      tags: [Device]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceRequest'
      responses:
        '201':
          description: Device registered and config push triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          description: Smart box or device type not found
  /api/devices/{id}:
    get:
      summary: Get device by ID
      operationId: getDevice
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          description: Device not found

    patch:
      summary: Update device properties
      operationId: updateDevice
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceRequest'
      responses:
        '200':
          description: Device updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          description: Device not found

    delete:
      summary: Delete a device
      operationId: deleteDevice
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Device deleted
        '404':
          description: Device not found
  /api/devices/{id}/command:
    post:
      summary: Send a command to a device (routed via MQTT to the SmartBox)
      operationId: sendDeviceCommand
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
        '400':
          description: Device is blocked
        '404':
          description: Device not found
  /api/devices/{id}/range:
    post:
      summary: Assign device to a range
      operationId: assignDeviceToRange
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRangeRequest'
      responses:
        '200':
          description: Device assigned to range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          description: Device or range not found

    delete:
      summary: Unassign device from its range
      operationId: removeDeviceFromRange
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Device unassigned
        '404':
          description: Device not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Range
  # ─────────────────────────────────────────────────────────────────────────────
...
  /api/ranges/{id}:
    get:
      summary: Get range by ID
      operationId: getRange
      tags: [Range]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Range found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeDetailResponse'
        '404':
          description: Range not found

    put:
      summary: Update a range
      operationId: updateRange
      tags: [Range]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRangeRequest'
      responses:
        '200':
          description: Range updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeSummaryResponse'
        '404':
          description: Range not found

    delete:
      summary: Delete a range
      operationId: deleteRange
      tags: [Range]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Range deleted
        '404':
          description: Range not found
...
  /api/smart-boxes/{id}:
    get:
      summary: Get smart box by ID
      operationId: getSmartBox
      tags: [SmartBox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Smart box found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartBoxResponse'
        '404':
          description: Smart box not found
...
components:
  schemas:
...
    DeviceTypeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Human-readable label (e.g. "Wurfmaschine Typ A")
        signalTypeId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        groupName:
          type: string
          description: Name of the device type group this type belongs to
        signalDurationMs:
          type: integer
          description: How long the signal is held active (ms)
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Optional delay before the signal fires (ms)
        command:
          type: string
          description: Command string sent to the firmware (e.g. ON, OFF)
        device:
          type: string
          description: Device kind as understood by the firmware (e.g. SENSOR, ACTUATOR)
        direction:
          type: string
          description: Communication direction (INPUT or OUTPUT)
...
    UpdateDeviceTypeRequest:
      type: object
      properties:
        name:
          type: string
          description: User-friendly label for this device type (e.g. "Wurfmaschine Typ A")
        signalDurationMs:
          type: integer
          minimum: 0
          nullable: true
          description: How long the signal is held active (ms)
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Optional delay before the signal fires (ms)

    # ───── Device ─────────────────────────────────────────────────────────────
    DeviceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alias:
          type: string
        smartBoxId:
          type: string
          format: uuid
          description: ID of the SmartBox this device is connected to
        groupId:
          type: string
          format: uuid
          description: Device type group
        groupName:
          type: string
          description: Name of the device type group
        deviceTypeId:
          type: string
          format: uuid
        deviceType:
          type: string
        command:
          type: string
          description: Firmware command (e.g. FIRE, ON)
        boxType:
          type: string
        blocked:
          type: boolean
        healthy:
          type: boolean
        signalDurationMs:
          type: integer
          description: How long the signal is held active
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Optional delay before signal fires
        pinConfig:
          $ref: '#/components/schemas/PinConfig'
        rangeId:
          type: string
          format: uuid
          nullable: true
...
    CommandResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
        detail:
          type: string

    # ───── Range ──────────────────────────────────────────────────────────────
    RangeSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        locked:
          type: boolean
          description: >
            true when the range is locked. Locked ranges reject device assignment changes.
            Use PATCH /api/ranges/{id}/locked to change this state.
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    RangeDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        locked:
          type: boolean
          description: >
            true when the range is locked. Locked ranges reject device assignment changes.
            Use PATCH /api/ranges/{id}/locked to change this state.
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceResponse'
        createdAt:
          type: string
          format: date-time
...
    UpdateRangeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...
    SmartBoxStatus:
      type: string
      enum: [ONLINE, OFFLINE, BLOCKED]
      description: SmartBox operational status

    # ───── Pin Configuration ───────────────────────────────────────────────────
    PinConfig:
      type: object
      required: [pin, mode]
      properties:
        pin:
          type: integer
          minimum: 0
          maximum: 40
          description: GPIO pin number on the Raspberry Pi Pico 2W (valid range 0–29 for user GPIOs; 0–40 reserved)
        mode:
          type: string
          enum: [OUTPUT, INPUT, INPUT_PULLUP, INPUT_PULLDOWN, PWM]
          description: GPIO mode as understood by the firmware
        pwmFrequencyHz:
          type: integer
          minimum: 1
          description: Required when mode is PWM. Sets the PWM carrier frequency.
        invertLogic:
          type: boolean
          default: false
          description: When true, the firmware inverts the signal (active-low wiring)

    # ───── Pagination ──────────────────────────────────────────────────────────
    PageMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    DevicePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DeviceResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    SmartBoxResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        macAddress:
          type: string
        alias:
          type: string
        status:
          $ref: '#/components/schemas/SmartBoxStatus'
        firmwareVersion:
          type: string
        configSynced:
          type: boolean
          description: >
            true once the SmartBox has ACK'd the latest config push via
            smartboxes/{mac}/config/ack
    CreateDeviceRequest:
      type: object
      required: [smartBoxId, groupId, deviceTypeId, alias]
      properties:
        smartBoxId:
          type: string
          format: uuid
          description: The SmartBox this device will be registered on
        groupId:
          type: string
          format: uuid
          description: The DeviceTypeGroup for this device
        deviceTypeId:
          type: string
          format: uuid
          description: The specific DeviceType within the group
        alias:
          type: string
          description: User-friendly name for this device
        rangeId:
          type: string
          format: uuid
          nullable: true
          description: Optional range to assign the device to on registration
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Override delay before the signal fires (ms)
    UpdateDeviceRequest:
      type: object
      properties:
        alias:
          type: string
        signalDurationMs:
          type: integer
          nullable: true
          description: How long the signal is held active (null = use device type default)
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Delay before signal fires (null = use device type default)
        groupId:
          type: string
          format: uuid
          description: Reassign device to a different device type group
    AssignRangeRequest:
      type: object
      required: [rangeId]
      properties:
        rangeId:
          type: string
          format: uuid
          description: ID of the range to assign this device to
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT